requirements.txt
# requirements.txt
#
# Core Google Cloud Libraries & APIs
# Using latest stable versions for better compatibility with Python 3.9+
google-cloud-bigquery>=3.20.0
google-cloud-aiplatform>=1.71.0 # Primary client for embeddings/models
# google-cloud-storage is generally useful but often a dependency of the others, keeping for safety
google-cloud-storage>=2.14.0 
google-cloud-secret-manager>=2.18.1

# Vertex AI and ML
# We use google-cloud-aiplatform for the client, and the standalone SDK for model calls.
vertexai>=1.71.0 # Latest stable version of the new SDK
google-generativeai==0.3.2 # For local quick-tests if needed, though we primarily use vertexai
tenacity==8.2.3  # Retry logic for API calls
cachetools==5.3.2  # Caching for embeddings/requests

# Data Processing
pandas==2.1.4
numpy==1.26.3
pyarrow==14.0.2

# Streamlit for UI
streamlit==1.30.0
streamlit-extras==0.3.6
streamlit-option-menu==0.3.6

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment & Utilities
python-dotenv==1.0.0
requests==2.31.0
httpx==0.26.0
python-dateutil==2.8.2
pytz==2023.3.post1
sqlalchemy==2.0.25
tqdm==4.66.1 # Added for the progress bars in your embedding script

# Testing (Development Dependencies)
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.23.3

# Code Quality (Development Dependencies)
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pre-commit==3.6.0
commitizen==3.13.0